Update number.ts
The allowNumeric option allows numerical strings to be parsed, but there is no case to handle strings with mixed content. 

Consider how these strings are currently parsed:
'123' -> 123
'abc123' -> Error
'123abc' -> 123

In the third example the string is parsed as a number. This behaviour is okay but it would be good to also have the option to be strict about the full numeracy of the string.

Enabling the fullNumberOnlyOption in conjunction with allowNumeric changes the mapping to this
'19' -> 19
'a19' -> Error
'19a' -> Error

Code untested.